合并商品的库存到商品列表接口一起查询

+++ src/main/java/com/jsh/erp/service/depotItem/DepotItemService.java
@@ -290,8 +290,8 @@ public class DepotItemService {
         return depotItemMapper.updateByPrimaryKeySelective(depotItem);
     }
 
-    public int findByTypeAndMaterialId(String type, Long mId) throws Exception{
-        int result =0;
+    public Long findByTypeAndMaterialId(String type, Long mId) throws Exception{
+        Long result = 0l;
         try{
             if(type.equals(TYPE)) {
                 result= depotItemMapperEx.findByTypeAndMaterialIdIn(mId);
